refactorisation for registration function
